+++ Player.h
@@ -10,13 +10,13 @@
 class Player
 {
 public:
-    Move get_move(Chess& ch, SearchLogger& search_log, int depth, U64& nodes, bool test);
+    Move get_move(Chess& ch, SearchLogger& search_log, int8_t depth, U64& nodes, bool test);
     Move get_book_move(Chess& ch, SearchLogger& search_log, bool test);
-    float iterative_search(Chess& ch, SearchLogger& search_log, int depth, bool test = false);
-    float nega_max(Chess& ch, SearchLogger& search_log, int depth, U64& nodes, float alpha = -99.99, float beta = 99.99, bool test = false);
-    float quiescence_search(Chess& ch, SearchLogger& search_log, int depth, U64& nodes, float alpha = -99.99, float beta = 99.99, bool test = false);
-    float eval(Chess& ch, std::vector<Move> moves, int mate_offset = 0, bool test = false);
-    float eval(Chess& ch, std::vector<Move> moves, SearchLogger& search_log, int mate_offset = 0, bool test = false);
+    float iterative_search(Chess& ch, SearchLogger& search_log, int8_t depth, bool test = false);
+    float nega_max(Chess& ch, SearchLogger& search_log, int8_t depth, U64& nodes, float alpha = -99.99, float beta = 99.99, bool test = false);
+    float quiescence_search(Chess& ch, SearchLogger& search_log, int8_t depth, U64& nodes, float alpha = -99.99, float beta = 99.99, bool test = false);
+    float eval(Chess& ch, std::vector<Move> moves, int8_t mate_offset = 0, bool test = false);
+    float eval(Chess& ch, std::vector<Move> moves, SearchLogger& search_log, int8_t mate_offset = 0, bool test = false);
     std::vector<Move> order_moves_by_piece(Chess& ch, std::vector<Move>);
 private:
     const float var_middlegame_weight = 32;
